+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "cyezoi" extension will be documented in this file.
 ## v0.0.31
 
 - **Add open in browser**
+- **Add copy problem markdown**
 - Fix some known issues
 - Fix record tree view auto refresh